chore: grid view record data table component
